requirements.txt
Django==4.1.7
django-environ==0.10.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-yasg==1.21.5
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
python-dotenv==1.0.0
pyTelegramBotAPI==4.12.0
python-telegram-bot==20.3
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
factory-boy==3.2.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
requests==2.31.0
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
pydantic==1.10.7
python-stdnum==1.18

